Move grpc services from server module to storage module
*Motivation*

Current almost every grpc requests are wrapped into `StorageContainerRequest` and their responses
are wrapped into `StorageContainerResponse`. It makes things a bit complicated on adding new grpc
services.

To simplify things, we can use grpc ClientInterceptor to stamp container information (e.g. scId)
into the request metadata and write a grpc service registry to take the container information from
request metadata and dispatch requests to containers.

In order to achieve it, we need to move the grpc services to storage container.

*Solution*

This PR moves grpc services from server module to storage module, so we can simplify the wire protocols.
