+++ stream/storage/impl/src/main/java/org/apache/bookkeeper/stream/storage/impl/grpc/GrpcTableService.java
@@ -15,15 +15,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.bookkeeper.stream.server.grpc;
+package org.apache.bookkeeper.stream.storage.impl.grpc;
 
 import io.grpc.stub.StreamObserver;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.bookkeeper.stream.proto.storage.StorageContainerRequest;
 import org.apache.bookkeeper.stream.proto.storage.StorageContainerResponse;
 import org.apache.bookkeeper.stream.proto.storage.TableServiceGrpc.TableServiceImplBase;
-import org.apache.bookkeeper.stream.server.handler.StorageContainerResponseHandler;
-import org.apache.bookkeeper.stream.storage.api.RangeStore;
+import org.apache.bookkeeper.stream.storage.impl.grpc.handler.StorageContainerResponseHandler;
+import org.apache.bookkeeper.stream.storage.api.metadata.RangeStoreService;
 
 /**
  * The gRPC protocol based k/v service.
@@ -31,9 +31,9 @@ import org.apache.bookkeeper.stream.storage.api.RangeStore;
 @Slf4j
 public class GrpcTableService extends TableServiceImplBase {
 
-    private final RangeStore rangeStore;
+    private final RangeStoreService rangeStore;
 
-    GrpcTableService(RangeStore store) {
+    public GrpcTableService(RangeStoreService store) {
         this.rangeStore = store;
         log.info("Created Table service");
     }
