+++ stream/storage/impl/src/main/java/org/apache/bookkeeper/stream/storage/impl/grpc/handler/ResponseHandler.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.bookkeeper.stream.server.handler;
+package org.apache.bookkeeper.stream.storage.impl.grpc.handler;
 
 import io.grpc.StatusException;
 import io.grpc.StatusRuntimeException;
